chore: Breakup monolithic Tokens/index.tsx into smaller components
