+++ app/components/UI/Tokens/TokenList/ScamWarningIcon/index.tsx
@@ -1,0 +1,48 @@
+import React from 'react';
+import { TokenI } from '../../types';
+import useIsOriginalNativeTokenSymbol from '../../../../hooks/useIsOriginalNativeTokenSymbol/useIsOriginalNativeTokenSymbol';
+import { useSelector } from 'react-redux';
+import {
+  selectChainId,
+  selectProviderConfig,
+  selectTicker,
+} from '../../../../../selectors/networkController';
+import ButtonIcon, {
+  ButtonIconSizes,
+} from '../../../../../../app/component-library/components/Buttons/ButtonIcon';
+import {
+  IconColor,
+  IconName,
+} from '../../../../../component-library/components/Icons/Icon';
+
+interface ScamWarningIconProps {
+  asset: TokenI;
+  setShowScamWarningModal: (arg: boolean) => void;
+}
+
+export const ScamWarningIcon = ({
+  asset,
+  setShowScamWarningModal,
+}: ScamWarningIconProps) => {
+  const { type } = useSelector(selectProviderConfig);
+  const chainId = useSelector(selectChainId);
+  const ticker = useSelector(selectTicker);
+  const isOriginalNativeTokenSymbol = useIsOriginalNativeTokenSymbol(
+    chainId,
+    ticker,
+    type,
+  );
+  if (!isOriginalNativeTokenSymbol && asset.isETH) {
+    return (
+      <ButtonIcon
+        iconName={IconName.Danger}
+        onPressIn={() => {
+          setShowScamWarningModal(true);
+        }}
+        iconColor={IconColor.Error}
+        size={ButtonIconSizes.Lg}
+      />
+    );
+  }
+  return null;
+};
